requirements.txt
pyTelegramBotAPI==4.19.0
Flask==3.0.3
python-dotenv==1.0.1
requests==2.32.3 # Optional, but good practice 